requirements.txt
anyio (2.2.0)
appdirs (1.4.4)
argon2-cffi (20.1.0)
asgiref (3.3.2)
async-generator (1.10)
attrs (20.3.0)
Babel (2.9.0)
backcall (0.2.0)
biopython (1.78)
bleach (3.3.0)
captcha (0.3)
certifi (2020.12.5)
cffi (1.14.5)
chardet (4.0.0)
crf-beam (0.0.1a0)
Cython (0.29.22)
datacache (1.1.5)
decorator (5.0.5)
defusedxml (0.7.1)
deprecation (2.1.0)
Django (3.1.7)
django-recaptcha (2.0.6)
entrypoints (0.3)
fast-ctc-decode (0.2.5)
gtfparse (1.2.1)
idna (2.10)
igv (0.10.0)
igv-jupyter (0.9.8)
install (1.3.4)
ipykernel (5.5.3)
ipython (7.22.0)
ipython-genutils (0.2.0)
ipywidgets (7.6.3)
jedi (0.18.0)
Jinja2 (2.11.3)
json5 (0.9.5)
jsonschema (3.2.0)
jupyter (1.0.0)
jupyter-client (6.2.0)
jupyter-console (6.4.0)
jupyter-core (4.7.1)
jupyter-packaging (0.9.1)
jupyter-server (1.6.1)
jupyterlab (3.0.14)
jupyterlab-pygments (0.1.2)
jupyterlab-server (2.4.0)
jupyterlab-widgets (1.0.0)
MarkupSafe (1.1.1)
memoized-property (1.0.3)
mistune (0.8.4)
mock (4.0.3)
nbclassic (0.2.7)
nbclient (0.5.3)
nbconvert (6.0.7)
nbformat (5.1.3)
nest-asyncio (1.5.1)
notebook (6.3.0)
numpy (1.18.5)
packaging (20.9)
pandas (1.2.4)
pandocfilters (1.4.3)
parso (0.8.2)
pexpect (4.8.0)
pickleshare (0.7.5)
Pillow (8.2.0)
pip (9.0.1)
pkg-resources (0.0.0)
progressbar33 (2.4)
prometheus-client (0.10.1)
prompt-toolkit (3.0.18)
ptyprocess (0.7.0)
pycparser (2.20)
pyensembl (1.9.1)
Pygments (2.8.1)
pyparsing (2.4.7)
pyrsistent (0.17.3)
python-dateutil (2.8.1)
pytz (2021.1)
PyYAML (5.4.1)
pyzmq (22.0.3)
qtconsole (5.0.3)
QtPy (1.9.0)
requests (2.25.1)
Send2Trash (1.5.0)
serializable (0.2.1)
setuptools (56.0.0)
simplejson (3.17.2)
six (1.15.0)
sniffio (1.2.0)
sqlparse (0.4.1)
terminado (0.9.4)
testpath (0.4.4)
tinytimer (0.0.0)
tomlkit (0.7.0)
tornado (6.1)
traitlets (5.0.5)
typechecks (0.1.0)
urllib3 (1.26.4)
wcwidth (0.2.5)
webencodings (0.5.1)
wheel (0.36.2)
widgetsnbextension (3.5.1)
